Update about.html
I marinated this Wagyu in fresh semen, and we're gonna dry age it

Starting with the semen, I squeeze my wife's belly after a day-long session of cuck fucking by random payed strangers on the street of Michigan. After placing all that fresh sperm in a container, I let the Wagyu marinate in the sauce for 24 hours.

After that, I brush some of my semen and place the Wagyu in the dry ager.

While I wait, I eat some Max Jerky Buffalo chicken ™️. It's actually my favorite flavor and you can find it at maxjerky.com

9 month laters and 2 rushes to the maternity ward, I take out the Wagyu.

The consistency was firm and the smell was amazing. I give it a pinch of pepper (as seminal juices already has a high quantity of salt in it), and we place it in the smoker.

We are roasting over Apple wood, low and slow

After 30 minutes, the results are incredible. The taste is firm and still has that semen aftertaste.